requirements.txt
gym[accept-rom-license, atari]==0.26.2
numpy==1.24.4
Pillow==10.1.0
torch==2.1.1
tqdm==4.66.1
